backend/requirements.txt
# Use flexible lower bounds so pip can install pre-built wheels on newer Python (e.g. 3.14).
# Fixed bcrypt compatibility for Python 3.14
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy>=2.0.36
pydantic>=2.10
pydantic-settings>=2.6
python-multipart>=0.0.20
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
